requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9